First pass at mvGLObject infrastructure.
mvGLObject is an extension of VRUI's GLObject system for separating
and managing application / context state. See the class documentation
for more information.

DataItem has been renamed mvContextState, and is in the process of
being refactored to only hold general shared context information, such
as the vtkRenderer, vtkLookupTable, etc. Specific actors/mappers/etc
will be moved into subclasses of mvGLObject and mvGLObject::DataItem.

Many of the MooseViewer ivars have been moved to the new
mvApplicationState class. An instance of this class is used to sync
the application state of the mvGLObject in MooseViewer::frame. This
holds relevant application state like the vtkReader and metadata
concerning the current scalar array. Other MooseViewer ivars have/will
be moved to mvGLObject subclasses.

The UnstructuredContourObject class has been converted to use the
mvGLObject framework as mvContours.