Add mvVolume for volume rendering.
Note that this currently has issues due to vtkCheckerboardSplatter
always acting on the active scalars rather than the specified input
array (SetInputArrayToProcess). There is a merge request pending to fix
this.

Started to rework the color handling code and eliminated some
calls to Vrui::requestUpdate() that happened every time
MooseViewer::frame() was called. Now the scene is only rerendered when
needed, rather than continuously.

Performance is much better now, as the old volume implementation was
doing multiple deep copies of the input dataset and re-splatting the
volume per-frame, per-context, whether or not volume rendering was used.
Volumes are now only calculated when the inputs change.