Add mvLODAsyncGLObject.
This new mvGLObject subclass allows three levels-of-detail to be used.
These are:

- Hint: Extremely fast update, usually performed synchronously so that
  *something* is always shown.
- LowRes: Computes a fast, low resolution visualization of the data.
  Eventually mvApplicationState will hold a low-res version of the
  data that can be used to implement this LOD.
- HiRes: Computes a high quality representation of the data from
  mvApplicationState::reader().dataObject().

If DataPipeline::forceSynchronousUpdates returns true, the update will
run in the UI thread. This is useful for hints. Otherwise, updates will
run in a background thread and go live once available. The highest
available LOD will always be rendered.

Each LOD is implemented by subclassing DataPipeline, LODData, and
RenderPipeline, which are returned from the virtual
mvLODAsyncGLObject::create* methods.